+++ code/Q3BSPFileParser.cpp
@@ -42,10 +42,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef ASSIMP_BUILD_NO_Q3BSP_IMPORTER
 
 #include "Q3BSPFileParser.h"
-#include "DefaultIOSystem.h"
 #include "Q3BSPFileData.h"
 #include "Q3BSPZipArchive.h"
 #include <vector>
+#include <assimp/DefaultIOSystem.h>
 #include <assimp/ai_assert.h>
 
 namespace Assimp {
