+++ include/assimp/DefaultIOStream.h
@@ -45,7 +45,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <stdio.h>
 #include <assimp/IOStream.hpp>
 #include <assimp/importerdesc.h>
-#include "Defines.h"
+#include <assimp/Defines.h>
 
 namespace Assimp    {
