+++ include/assimp/Defines.h
