+++ test/unit/TestIOStream.h
@@ -39,7 +39,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 #pragma once
 
-#include "DefaultIOStream.h"
+#include <assimp/DefaultIOStream.h>
 
 using namespace ::Assimp;
