+++ test/unit/utImporter.cpp
@@ -45,7 +45,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <assimp/Importer.hpp>
 #include <BaseImporter.h>
 #include "TestIOSystem.h"
-#include "DefaultIOSystem.h"
+#include <assimp/DefaultIOSystem.h>
 
 using namespace ::std;
 using namespace ::Assimp;
@@ -53,12 +53,12 @@ using namespace ::Assimp;
 class ImporterTest : public ::testing::Test
 {
 public:
-    virtual void SetUp() { 
-        pImp = new Importer(); 
+    virtual void SetUp() {
+        pImp = new Importer();
     }
-    
-    virtual void TearDown() { 
-        delete pImp; 
+
+    virtual void TearDown() {
+        delete pImp;
     }
 
 protected:
